image/requirements.txt
Flask
requests
python-dotenv
paho-mqtt
